+++ src/main/java/com/fourthread/ozang/module/domain/security/jwt/handler/JwtLogoutHandler.java
@@ -1,9 +1,9 @@
-package com.fourthread.ozang.module.domain.security.jwt;
+package com.fourthread.ozang.module.domain.security.jwt.handler;
 
+import com.fourthread.ozang.module.domain.security.jwt.JwtService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.time.Instant;
 import java.util.Arrays;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
