+++ src/main/java/com/fourthread/ozang/module/domain/security/oauth/parser/KakaoOAuthAttributesParser.java
@@ -1,6 +1,6 @@
-package com.fourthread.ozang.module.domain.oauth.parser;
+package com.fourthread.ozang.module.domain.security.oauth.parser;
 
-import com.fourthread.ozang.module.domain.oauth.dto.OAuthAttributes;
+import com.fourthread.ozang.module.domain.security.oauth.dto.OAuthAttributes;
 import com.fourthread.ozang.module.domain.user.dto.type.Role;
 import java.util.Map;
 import org.springframework.stereotype.Component;
